extracted the dns monitor to its own monitor-type
