+++ server/model/monitor.js
@@ -481,46 +481,6 @@ class Monitor extends BeanModel {
                     bean.ping = await ping(this.hostname, this.packetSize);
                     bean.msg = "";
                     bean.status = UP;
-                } else if (this.type === "dns") {
-                    let startTime = dayjs().valueOf();
-                    let dnsMessage = "";
-
-                    let dnsRes = await dnsResolve(this.hostname, this.dns_resolve_server, this.port, this.dns_resolve_type);
-                    bean.ping = dayjs().valueOf() - startTime;
-
-                    if (this.dns_resolve_type === "A" || this.dns_resolve_type === "AAAA" || this.dns_resolve_type === "TXT") {
-                        dnsMessage += "Records: ";
-                        dnsMessage += dnsRes.join(" | ");
-                    } else if (this.dns_resolve_type === "CNAME" || this.dns_resolve_type === "PTR") {
-                        dnsMessage = dnsRes[0];
-                    } else if (this.dns_resolve_type === "CAA") {
-                        dnsMessage = dnsRes[0].issue;
-                    } else if (this.dns_resolve_type === "MX") {
-                        dnsRes.forEach(record => {
-                            dnsMessage += `Hostname: ${record.exchange} - Priority: ${record.priority} | `;
-                        });
-                        dnsMessage = dnsMessage.slice(0, -2);
-                    } else if (this.dns_resolve_type === "NS") {
-                        dnsMessage += "Servers: ";
-                        dnsMessage += dnsRes.join(" | ");
-                    } else if (this.dns_resolve_type === "SOA") {
-                        dnsMessage += `NS-Name: ${dnsRes.nsname} | Hostmaster: ${dnsRes.hostmaster} | Serial: ${dnsRes.serial} | Refresh: ${dnsRes.refresh} | Retry: ${dnsRes.retry} | Expire: ${dnsRes.expire} | MinTTL: ${dnsRes.minttl}`;
-                    } else if (this.dns_resolve_type === "SRV") {
-                        dnsRes.forEach(record => {
-                            dnsMessage += `Name: ${record.name} | Port: ${record.port} | Priority: ${record.priority} | Weight: ${record.weight} | `;
-                        });
-                        dnsMessage = dnsMessage.slice(0, -2);
-                    }
-
-                    if (this.dnsLastResult !== dnsMessage) {
-                        R.exec("UPDATE `monitor` SET dns_last_result = ? WHERE id = ? ", [
-                            dnsMessage,
-                            this.id
-                        ]);
-                    }
-
-                    bean.msg = dnsMessage;
-                    bean.status = UP;
                 } else if (this.type === "push") {      // Type: Push
                     log.debug("monitor", `[${this.name}] Checking monitor at ${dayjs().format("YYYY-MM-DD HH:mm:ss.SSS")}`);
                     const bufferTime = 1000; // 1s buffer to accommodate clock differences
